+++ Assets/Plugins/other/QFramework/Toolkits/ResKit/Scripts/HotFixDowload/HotFixScripts/MD5Manager.cs
@@ -1,30 +1,31 @@
-using UnityEngine;
-using System.Collections;
-using System.Collections.Generic;
 using System;
-using System.Security.Cryptography;
 using System.IO;
+using System.Security.Cryptography;
 using System.Text;
-
+using UnityEngine;
 
 namespace QFramework
 {
     [Serializable]
-    public class MD5Manager  : ISingleton
+    public class MD5Manager : ISingleton
     {
+        private MD5Manager()
+        {
+        }
+
         public static MD5Manager Instance => SingletonProperty<MD5Manager>.Instance;
-        private MD5Manager() {}
+
+        public void OnSingletonInit()
+        {
+        }
 
         //储存Md5码，filePath为文件路径，md5SavePath为储存md5码路径
         public void SaveMd5(string filePath, string md5SavePath)
         {
-            string md5 = BuildFileMd5(filePath);
-            string name = filePath + "_md5.dat";
-            if (File.Exists(name))
-            {
-                File.Delete(name);
-            }
-            StreamWriter sw = new StreamWriter(name, false, Encoding.UTF8);
+            var md5 = BuildFileMd5(filePath);
+            var name = filePath + "_md5.dat";
+            if (File.Exists(name)) File.Delete(name);
+            var sw = new StreamWriter(name, false, Encoding.UTF8);
             if (sw != null)
             {
                 sw.Write(md5);
@@ -36,13 +37,10 @@ namespace QFramework
         //储存Md5码，filePath为文件路径
         public void SaveMd5(string filePath)
         {
-            string md5 = BuildFileMd5(filePath);
-            string name = filePath + "_md5.dat";
-            if (File.Exists(name))
-            {
-                File.Delete(name);
-            }
-            StreamWriter sw = new StreamWriter(name, false, Encoding.UTF8);
+            var md5 = BuildFileMd5(filePath);
+            var name = filePath + "_md5.dat";
+            if (File.Exists(name)) File.Delete(name);
+            var sw = new StreamWriter(name, false, Encoding.UTF8);
             if (sw != null)
             {
                 sw.Write(md5);
@@ -54,11 +52,11 @@ namespace QFramework
         //获取之前储存的Md5码
         public string GetMd5(string path)
         {
-            string name = path + "_md5.dat";
+            var name = path + "_md5.dat";
             try
             {
-                StreamReader sr = new StreamReader(name, Encoding.UTF8);
-                string content = sr.ReadToEnd();
+                var sr = new StreamReader(name, Encoding.UTF8);
+                var content = sr.ReadToEnd();
                 sr.Close();
                 return content;
             }
@@ -76,27 +74,22 @@ namespace QFramework
                 using (var fileStream = File.OpenRead(fliePath))
                 {
                     var md5 = MD5.Create();
-                    var fileMD5Bytes = md5.ComputeHash(fileStream);//计算指定Stream 对象的哈希值                                     
+                    var fileMD5Bytes =
+                        md5.ComputeHash(fileStream); //计算指定Stream 对象的哈希值                                     
                     filemd5 = FormatMD5(fileMD5Bytes);
-
                 }
             }
-            catch (System.Exception ex)
+            catch (Exception ex)
             {
                 Debug.LogError(ex);
             }
+
             return filemd5;
         }
 
-        public string FormatMD5(Byte[] data)
+        public string FormatMD5(byte[] data)
         {
-            return System.BitConverter.ToString(data).Replace("-", "").ToLower();//将byte[]装换成字符串
-        }
-
-        public void OnSingletonInit()
-        {
-            
+            return BitConverter.ToString(data).Replace("-", "").ToLower(); //将byte[]装换成字符串
         }
     }
 }
-
